+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 
 - Viewport Height component no longer sets a `height` except for IE  
+- Viewport Height component no longer forces `box-sizing: border-box`  
 - Improve Tooltip performance
 - Improve Parallax performance